+++ activerecord/lib/active_record/log_subscriber.rb
@@ -6,27 +6,6 @@ module ActiveRecord
 
     class_attribute :backtrace_cleaner, default: ActiveSupport::BacktraceCleaner.new
 
-    def self.runtime=(value)
-      ActiveRecord.deprecator.warn(<<-MSG.squish)
-        ActiveRecord::LogSubscriber.runtime= is deprecated and will be removed in Rails 7.2.
-      MSG
-      ActiveRecord::RuntimeRegistry.sql_runtime = value
-    end
-
-    def self.runtime
-      ActiveRecord.deprecator.warn(<<-MSG.squish)
-        ActiveRecord::LogSubscriber.runtime is deprecated and will be removed in Rails 7.2.
-      MSG
-      ActiveRecord::RuntimeRegistry.sql_runtime
-    end
-
-    def self.reset_runtime
-      ActiveRecord.deprecator.warn(<<-MSG.squish)
-        ActiveRecord::LogSubscriber.reset_runtime is deprecated and will be removed in Rails 7.2.
-      MSG
-      ActiveRecord::RuntimeRegistry.reset
-    end
-
     def strict_loading_violation(event)
       debug do
         owner = event.payload[:owner]